# platform/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_attendance.submitRequest:
    properties:
      end_time:
        description: RFC3339
        type: string
      reason:
        type: string
      start_time:
        description: RFC3339
        type: string
      status:
        description: rest/overtime/leave
        type: string
      supervisor_id:
        description: 直接上级，提交后转其审批
        type: string
    required:
    - end_time
    - start_time
    - status
    - supervisor_id
    type: object
...
info:
  contact: {}
  description: 企业 AI 原生研发平台后端（AI 驱动 需求→研发→测试→审批→发布 全流程）
  title: 智源 ANP 平台 API
  version: "1.0"
paths:
...
  /attendance/{id}/approve:
    post:
      parameters:
      - description: 考勤记录ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 审批后的考勤记录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 审批通过考勤
      tags:
      - attendance
  /attendance/{id}/reject:
    post:
      parameters:
      - description: 考勤记录ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 驳回后的考勤记录
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 驳回考勤
      tags:
      - attendance
  /attendance/inbox:
    get:
      parameters:
      - description: 状态过滤(pending/approved/rejected)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 待审批考勤列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 待我审批的考勤收件箱
      tags:
      - attendance
...
  /project-spaces/{id}/attendance:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 考勤记录列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 列出项目空间下的考勤记录
      tags:
      - attendance
    post:
      consumes:
      - application/json
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      - description: 考勤提交内容(status+start_time+end_time+supervisor_id)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_attendance.submitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的考勤记录
          schema:
            additionalProperties: true
            type: object
        "400":
          description: invalid body / 时间格式错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 提交考勤记录
      tags:
      - attendance
  /project-spaces/{id}/attendance/mine:
    get:
      parameters:
      - description: 项目空间ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 当前用户的考勤记录列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务端错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 当前员工的考勤记录
      tags:
      - attendance
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"